requirements.txt
streamlit
pandas
openpyxl
google-generativeai
playwright
beautifulsoup4